LayoutTests/fast/viewport/viewport-31-expected.txt

CONSOLE MESSAGE: line 3: Viewport height or height set to physical device height, try using "device-height" constant instead for future compatibility.
viewport size 980x700 scale 0.502857 with limits [0.502857, 5.000000]
